# server/openapi.json
{
  "x-generator": "NSwag v14.6.3.0 (NJsonSchema v11.5.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_RegisterUser",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterLoginDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_LoginUser",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterLoginDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/realtime/connect": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Connect",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/realtime/join": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_JoinGroup",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGroupRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinGroupBroadcast"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinGroupResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLeftResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/realtime/createroom": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_CreateChatRoom",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatRoom"
                }
              }
            }
          }
        }
      }
    },
    "/api/realtime/getmembers": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_GetGroupMembers",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/realtime/getonlinemembers": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_GetOnlineGroupMembers",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/realtime/GetRooms": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_GetRooms",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatRoom"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/realtime/send": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Send",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendGroupMessageRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/realtime/poke": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Poke",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PokeRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokeResponseDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "userId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "passwordHash": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "chatRooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserChatRoom"
            }
          }
        }
      },
      "Message": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "chatRoomId": {
            "type": "string"
          },
          "chatRoom": {
            "$ref": "#/components/schemas/ChatRoom"
          }
        }
      },
      "ChatRoom": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserChatRoom"
            }
          }
        }
      },
      "UserChatRoom": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "chatRoomId": {
            "type": "string"
          },
          "chatRoom": {
            "$ref": "#/components/schemas/ChatRoom"
          }
        }
      },
      "RegisterLoginDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "userName",
          "password"
        ],
        "properties": {
          "userName": {
            "type": "string",
            "minLength": 3
          },
          "password": {
            "type": "string",
            "minLength": 6
          }
        }
      },
      "LoginUserDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "ConnectionResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "connectionId": {
                "type": "string"
              }
            }
          }
        ]
      },
      "BaseResponseDto": {
        "type": "object",
        "description": "Models used for server sent events to clients should preferably use this model.\nIt automatically attaches \"EventType\" as a property with type name as value.",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "eventType": {
            "type": "string"
          }
        }
      },
      "JoinGroupBroadcast": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "connectedUsers": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ConnectionIdAndUserName"
                }
              }
            }
          }
        ]
      },
      "ConnectionIdAndUserName": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "connectionId": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          }
        }
      },
      "JoinGroupResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "chatroom": {
                "$ref": "#/components/schemas/ChatRoom"
              }
            }
          }
        ]
      },
      "UserLeftResponseDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "connectionId": {
                "type": "string"
              }
            }
          }
        ]
      },
      "JoinGroupRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "connectionId": {
            "type": "string"
          },
          "group": {
            "type": "string"
          }
        }
      },
      "MessageResponseDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "message": {
                "type": "string"
              },
              "user": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "SendGroupMessageRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          }
        }
      },
      "PokeResponseDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "pokedBy": {
                "type": "string"
              }
            }
          }
        ]
      },
      "PokeRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "connectionIdToPoke": {
            "type": "string"
          }
        }
      }
    }
  }
}